Feature detail

+++ routes/character_detail.js
@@ -31,12 +31,12 @@ router.get("/", async (req, res, next) => {
 });
 
 router.post("/write_review", async (req, res, next) => {
-  //리뷰 작성
+  //리뷰 작성하기
   const {user_id, character_id, review_content, review_spoiler} = req.body;
   let result_code = 404;
   let message = "에러가 발생했습니다.";
   try {
-    await pool.execute(
+    let [insert_review] = await pool.execute(
       `INSERT INTO character_review(user_id, character_id, review_content, review_spoiler) VALUES (?, ?, ?, ?)`,
       [user_id, character_id, review_content, review_spoiler]
     );
@@ -47,6 +47,7 @@ router.post("/write_review", async (req, res, next) => {
       message: message,
       user_id: user_id,
       character_id: character_id,
+      review_id: insert_review.insertId,
     });
   } catch (err) {
     console.error(err);
@@ -124,15 +125,49 @@ router.post("/review_dislike", async (req, res, next) => {
   }
 });
 
+router.post("/modify_review", async (req, res, next) => {
+  //리뷰 수정하기
+  const {user_id, character_id, review_content, review_spoiler, review_id} =
+    req.body;
+  let result_code = 404;
+  let message = "에러가 발생했습니다.";
+  try {
+    await pool.execute(
+      `UPDATE character_review SET review_content = ?, review_spoiler = ? WHERE review_id = ?`,
+      [review_content, review_spoiler, review_id]
+    );
+
+    result_code = 200;
+    message = "리뷰를 수정했습니다.";
+
+    return res.json({
+      code: result_code,
+      message: message,
+      user_id: user_id,
+      character_id: character_id,
+      review_id: review_id,
+    });
+  } catch (err) {
+    console.error(err);
+    return res.status(500).json(error);
+  }
+});
+
 router.delete("/delete_review", async (req, res, next) => {
   //리뷰 삭제하기
   const {user_id, character_id, review_id} = req.body;
   let result_code = 404;
   let message = "에러가 발생했습니다.";
   try {
-    await pool.execute(`DELETE FROM review_like WHERE review_id = ?`, [
-      review_id,
-    ]);
+    let [is_review_exist] = await pool.execute(
+      `SELECT * FROM review_like WHERE review_id = ?`,
+      [review_id]
+    );
+    if (is_review_exist[0] != null) {
+      await pool.execute(`DELETE FROM review_like WHERE review_id = ?`, [
+        review_id,
+      ]);
+    }
     await pool.execute(`DELETE FROM character_review WHERE review_id = ?`, [
       review_id,
     ]);
@@ -144,6 +179,118 @@ router.delete("/delete_review", async (req, res, next) => {
       message: message,
       user_id: user_id,
       character_id: character_id,
+    });
+  } catch (err) {
+    console.error(err);
+    return res.status(500).json(error);
+  }
+});
+
+router.get("/comment", async (req, res, next) => {
+  //댓글 조회하기
+  const {user_id, review_id} = req.query;
+  let result_code = 404;
+  let message = "에러가 발생했습니다.";
+  try {
+    let [view_comment] = await pool.execute(
+      `SELECT * FROM review_comment WHERE review_id = ?`,
+      [review_id]
+    );
+    result_code = 200;
+    message = "리뷰를 조회했습니다.";
+    return res.json({
+      code: result_code,
+      message: message,
+      user_id: user_id,
+      comment_list: view_comment,
+    });
+  } catch (err) {
+    console.error(err);
+    return res.status(500).json(error);
+  }
+});
+
+router.post("/write_comment", async (req, res, next) => {
+  //댓글 쓰기
+  const {user_id, character_id, review_id, comment_content, comment_spoiler} =
+    req.body;
+  let result_code = 404;
+  let message = "에러가 발생했습니다.";
+  try {
+    let [insert_comment] = await pool.execute(
+      `INSERT INTO review_comment (review_id, user_id, comment_content, comment_spoiler, character_id) VALUES (?, ?, ?, ?, ?)`,
+      [review_id, user_id, comment_content, comment_spoiler, character_id]
+    );
+    result_code = 200;
+    message = "댓글을 작성했습니다.";
+    return res.json({
+      code: result_code,
+      message: message,
+      user_id: user_id,
+      character_id: character_id,
+      comment_id: insert_comment.insertId,
+    });
+  } catch (err) {
+    console.error(err);
+    return res.status(500).json(error);
+  }
+});
+
+router.post("/modify_comment", async (req, res, next) => {
+  //댓글 수정하기
+  const {user_id, character_id, comment_content, comment_spoiler, comment_id} =
+    req.body;
+  let result_code = 404;
+  let message = "에러가 발생했습니다.";
+  try {
+    await pool.execute(
+      `UPDATE review_comment SET comment_content = ?, comment_spoiler = ? WHERE comment_id = ?`,
+      [comment_content, comment_spoiler, comment_id]
+    );
+
+    result_code = 200;
+    message = "댓글을 수정했습니다.";
+
+    return res.json({
+      code: result_code,
+      message: message,
+      user_id: user_id,
+      character_id: character_id,
+      comment_id: comment_id,
+    });
+  } catch (err) {
+    console.error(err);
+    return res.status(500).json(error);
+  }
+});
+
+router.delete("/delete_comment", async (req, res, next) => {
+  //댓글 삭제하기
+  const {user_id, character_id, review_id, comment_id} = req.body;
+  let result_code = 404;
+  let message = "에러가 발생했습니다.";
+  try {
+    let [is_comment_exist] = await pool.execute(
+      `SELECT * FROM comment_like WHERE comment_id = ?`,
+      [comment_id]
+    );
+    if (is_comment_exist[0] != null) {
+      await pool.execute(`DELETE FROM comment_like WHERE comment_id = ?`, [
+        comment_id,
+      ]);
+    }
+    await pool.execute(`DELETE FROM review_comment WHERE comment_id = ?`, [
+      comment_id,
+    ]);
+    result_code = 200;
+    message = "댓글을 삭제했습니다.";
+
+    return res.json({
+      code: result_code,
+      message: message,
+      user_id: user_id,
+      character_id: character_id,
+      review_id: review_id,
     });
   } catch (err) {
     console.error(err);
